refactor: move about compose ui
- keep about activity focused on lifecycle setup
